Finalise refactoring of network.py. Resolve #11
