+++ models/autoencoders/conditionalVAE.py
@@ -6,7 +6,7 @@ Author: Eric Drechsler (eric_drechsler@sfu.ca)
 """
 
 import torch
-from models.models.variationalAE import VariationalAutoEncoder
+from models.autoencoders.variationalAE import VariationalAutoEncoder
 
 #logging module with handmade settings.
 from DiVAE import logging
