+++ models/autoencoders/variationalAE.py
@@ -6,7 +6,7 @@ Author: Eric Drechsler (eric_drechsler@sfu.ca)
 """
 import torch
 from torch import nn
-from models.models.autoencoder import AutoEncoder
+from models.autoencoders.autoencoder import AutoEncoder
 
 #logging module with handmade settings.
 from DiVAE import logging
